yago/requirements.txt
# YAGO Dependencies

# Core AI Frameworks
crewai>=0.28.0
crewai-tools>=0.2.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0

# API Clients
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0

# File & Data Processing
gitpython>=3.1.0

# Testing (for generated code)
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
ruff>=0.1.0
black>=23.0.0

# Web Dashboard
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
